Add files via upload
This is some old code I made a long time ago. I will go back and debug it since it broke over time cause you know.... web scraping. This code is not the most reproducible since it logs to a database which is on my localhost. I will see to make some changes to it so that it simply logs to a file.